# openapi/state-v1.yaml
openapi: 3.1.0
info:
  title: State API
  version: 1.0.0
  description: REST contract for the self-hosted State reminder and context service.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: https://state.example.com
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/notes/{note_id}/attachments:
    parameters:
      - $ref: "#/components/parameters/NoteID"
    post:
      tags: [Notes]
      operationId: uploadNoteAttachment
      description: Stores one photo or recording as the raw request body. The server checks type, signature, size and the SHA-256 the client sent before attaching it. Runners are rejected.
      parameters:
        - { name: X-State-Client-Request-Id, in: header, required: true, schema: { type: string } }
        - { name: X-State-Ordinal, in: header, required: true, schema: { type: integer, minimum: 0 } }
        - { name: X-State-Content-SHA256, in: header, required: true, schema: { type: string, pattern: "^[0-9a-f]{64}$" } }
        - { name: X-State-Duration-Ms, in: header, required: false, schema: { type: integer, minimum: 0 } }
      requestBody:
        required: true
        content:
          image/jpeg: { schema: { type: string, format: binary } }
          image/png: { schema: { type: string, format: binary } }
          image/heic: { schema: { type: string, format: binary } }
          audio/mp4: { schema: { type: string, format: binary } }
          audio/m4a: { schema: { type: string, format: binary } }
          audio/aac: { schema: { type: string, format: binary } }
          audio/mpeg: { schema: { type: string, format: binary } }
          audio/wav: { schema: { type: string, format: binary } }
      responses:
        "201":
          description: Attachment stored.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NoteView" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/notes/{note_id}/attachments/{attachment_id}:
    parameters:
      - $ref: "#/components/parameters/NoteID"
      - { name: attachment_id, in: path, required: true, schema: { $ref: "#/components/schemas/UUIDv7" } }
    get:
      tags: [Notes]
      operationId: downloadNoteAttachment
      description: Streams the original photo or recording.
      responses:
        "200":
          description: The original bytes with their MIME type.
          content:
            application/octet-stream: { schema: { type: string, format: binary } }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/notes/{note_id}/processing:
    parameters:
      - $ref: "#/components/parameters/NoteID"
    post:
      tags: [Notes]
      operationId: requestNoteProcessing
      description: Asks the notes AI to transcribe, read and organize the note. Without an OpenRouter key on the server or without the owner's consent nothing is sent and the status says why.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [client_request_id]
              properties:
                client_request_id: { type: string }
      responses:
        "202":
          description: Processing state recorded.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NoteView" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    get:
      tags: [Notes]
      operationId: getNoteProcessing
      responses:
        "200":
          description: Processing state, AI result and attachments with their OCR or transcript.
          content:
            application/json:
              schema:
                type: object
                properties:
                  processing: { $ref: "#/components/schemas/NoteProcessing" }
                  ai: { $ref: "#/components/schemas/NoteAIResult" }
                  attachments:
                    type: array
                    items: { $ref: "#/components/schemas/NoteAttachment" }
  /api/v1/notes/{note_id}/relations:
    parameters:
      - $ref: "#/components/parameters/NoteID"
    get:
      tags: [Notes]
      operationId: listNoteRelations
      responses:
        "200":
          description: Related notes the AI linked, with the reason.
          content:
            application/json:
              schema:
                type: object
                properties:
                  relations:
                    type: array
                    items: { $ref: "#/components/schemas/NoteRelation" }
  /api/v1/notes/{note_id}/relations/{relation_id}:
    parameters:
      - $ref: "#/components/parameters/NoteID"
      - { name: relation_id, in: path, required: true, schema: { $ref: "#/components/schemas/UUIDv7" } }
    patch:
      tags: [Notes]
      operationId: dismissNoteRelation
      description: Owner and devices only. Archives a link; nothing is deleted.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [archived]
              properties:
                archived: { type: boolean, enum: [true] }
      responses:
        "200":
          description: Note after the change.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NoteView" }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/notes/{note_id}/reminder-proposals/{proposal_id}/accept:
    parameters:
      - $ref: "#/components/parameters/NoteID"
      - { name: proposal_id, in: path, required: true, schema: { $ref: "#/components/schemas/UUIDv7" } }
    post:
      tags: [Notes]
      operationId: acceptReminderProposal
      description: Owner and devices only. Creates the proposed reminder; an agent's inference alone never creates one.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [client_request_id]
              properties:
                time_zone: { type: string, description: IANA zone, required when the proposal has a date. }
                client_request_id: { type: string }
      responses:
        "201":
          description: Reminder created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Reminder" }
        "403": { $ref: "#/components/responses/Error" }
  /api/v1/notes/{note_id}/reminder-proposals/{proposal_id}/dismiss:
    parameters:
      - $ref: "#/components/parameters/NoteID"
      - { name: proposal_id, in: path, required: true, schema: { $ref: "#/components/schemas/UUIDv7" } }
    post:
      tags: [Notes]
      operationId: dismissReminderProposal
      responses:
        "200":
          description: Note after the change.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NoteView" }
        "403": { $ref: "#/components/responses/Error" }
...
  /api/v1/notes-ai/settings:
    get:
      tags: [Notes]
      operationId: getNotesAISettings
      responses:
        "200":
          description: Consent, monthly limit, models and this month's spending. Never the key.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotesAISettings" }
    patch:
      tags: [Notes]
      operationId: updateNotesAISettings
      description: Owner and devices only.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                consent: { type: boolean }
                monthly_limit_usd: { type: number, exclusiveMinimum: 0, maximum: 1000 }
                agent_model: { type: string }
                transcription_model: { type: string }
      responses:
        "200":
          description: Settings after the change.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotesAISettings" }
        "403": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    NoteID:
      name: note_id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/UUIDv7" }
...
  responses:
    Error:
      description: Structured API error.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    UUIDv7:
      type: string
      format: uuid
      pattern: "^[0-9a-f]{8}-[0-9a-f]{4}-7[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
...
    ReminderStatus:
      type: string
      enum: [active, completed]
    TimeZoneMode:
      type: string
      enum: [floating, fixed]
    Schedule:
      type: object
      required: [local_date, time_zone, mode]
      properties:
        local_date: { type: string, format: date }
        local_time:
          type: [string, "null"]
          pattern: "^[0-2][0-9]:[0-5][0-9]$"
        time_zone:
          type: string
          description: IANA time zone identifier.
        mode: { $ref: "#/components/schemas/TimeZoneMode" }
        prewarning_minutes: { type: [integer, "null"], minimum: 0, maximum: 525600 }
    RecurrenceRule:
      type: object
      required: [frequency, interval]
      properties:
        frequency: { type: string, enum: [daily, weekly, monthly, yearly] }
        interval: { type: integer, minimum: 1, maximum: 100 }
        until_date: { type: [string, "null"], format: date }
    Reminder:
      type: object
      required: [id, title, status, revision, archived, created_at, updated_at]
      properties:
        id: { $ref: "#/components/schemas/UUIDv7" }
        title: { type: string, minLength: 1, maxLength: 500 }
        description: { type: [string, "null"], maxLength: 100000 }
        status: { $ref: "#/components/schemas/ReminderStatus" }
        schedule:
          oneOf:
            - $ref: "#/components/schemas/Schedule"
            - type: "null"
        recurrence:
          oneOf:
            - $ref: "#/components/schemas/RecurrenceRule"
            - type: "null"
        execution_policy_id:
          oneOf:
            - $ref: "#/components/schemas/UUIDv7"
            - type: "null"
        revision: { type: integer, format: int64, minimum: 1 }
        archived: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    NoteFieldSource:
      type: string
      enum: [user, derived, ai]
      description: user values are never overwritten automatically; derived values follow the document; ai values come from the notes AI and never replace a user value.
    Note:
      type: object
      required: [id, title, title_source, document, plain_text, summary, summary_source, archived, revision, created_at, updated_at]
      properties:
        id: { $ref: "#/components/schemas/UUIDv7" }
        title: { type: string, maxLength: 200 }
        title_source: { $ref: "#/components/schemas/NoteFieldSource" }
        document:
          type: string
          description: Canonical content as Markdown (CommonMark plus - [ ] checklists and --- dividers). At most 262144 UTF-8 bytes; JSON Schema maxLength counts characters, so the byte limit is enforced by the server.
        plain_text: { type: string, description: Derived text without Markdown markers, used for search. }
        summary: { type: string, maxLength: 280 }
        summary_source: { $ref: "#/components/schemas/NoteFieldSource" }
        capture: { type: string, enum: [image, audio], description: Set for photo and voice notes, which may start without text. }
        archived: { type: boolean }
        revision: { type: integer, format: int64, minimum: 1 }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    NoteView:
      description: A note as clients see it. An AI title or summary replaces only a derived one.
      allOf:
        - $ref: "#/components/schemas/Note"
        - type: object
          required: [attachments, processing, relations, reminder_proposals]
          properties:
            attachments:
              type: array
              items: { $ref: "#/components/schemas/NoteAttachment" }
            processing: { $ref: "#/components/schemas/NoteProcessing" }
            ai: { $ref: "#/components/schemas/NoteAIResult" }
            relations:
              type: array
              items: { $ref: "#/components/schemas/NoteRelation" }
            reminder_proposals:
              type: array
              items: { $ref: "#/components/schemas/ReminderProposal" }
    NoteAttachment:
      type: object
      required: [id, note_id, ordinal, kind, mime_type, byte_size, sha256, created_at]
      properties:
        id: { $ref: "#/components/schemas/UUIDv7" }
        note_id: { $ref: "#/components/schemas/UUIDv7" }
        ordinal: { type: integer, minimum: 0 }
        kind: { type: string, enum: [image, audio] }
        mime_type: { type: string }
        byte_size: { type: integer, format: int64 }
        sha256: { type: string }
        duration_ms: { type: integer, format: int64 }
        derived_text: { type: string, description: OCR text of a photo or transcript of a recording. }
        derived_kind: { type: string, enum: [ocr, transcript] }
        derived_model: { type: string }
        created_at: { type: string, format: date-time }
    NoteProcessing:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [idle, queued, processing, ready, needs_review, failed, consent_required, not_configured, budget_exhausted] }
        error: { type: string }
        model: { type: string }
        updated_at: { type: string, format: date-time }
    NoteAIResult:
      type: object
      properties:
        title: { type: string }
        summary: { type: string }
        model: { type: string }
        source_hash: { type: string, description: SHA-256 of the plain text the result was made from. }
        proposed_document: { type: string, description: Body the AI suggests when the owner already wrote text. }
        needs_review: { type: boolean }
        updated_at: { type: string, format: date-time }
    NoteRelation:
      type: object
      properties:
        id: { $ref: "#/components/schemas/UUIDv7" }
        note_id: { $ref: "#/components/schemas/UUIDv7" }
        related_note_id: { $ref: "#/components/schemas/UUIDv7" }
        related_title: { type: string }
        reason: { type: string }
        confidence: { type: number }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
    ReminderProposal:
      type: object
      properties:
        id: { $ref: "#/components/schemas/UUIDv7" }
        note_id: { $ref: "#/components/schemas/UUIDv7" }
        title: { type: string }
        description: { type: string }
        local_date: { type: string }
        local_time: { type: string }
        reason: { type: string }
        status: { type: string, enum: [pending, accepted, dismissed] }
        reminder_id: { type: string }
...
    NotesAISettings:
      type: object
      properties:
        consent: { type: boolean }
        consent_at: { type: string, format: date-time }
        monthly_limit_usd: { type: number }
        agent_model: { type: string }
        transcription_model: { type: string }
        month: { type: string }
        spent_this_month_usd: { type: number }
        key_configured: { type: boolean }
...
    Error:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: [object, array, string, number, boolean, "null"] }